Authentication for API done
