+++ cat/pages/api/dbMakerAPI/cartas/editField.ts
@@ -2,7 +2,7 @@ import { NextApiRequest, NextApiResponse } from "next"
 import { DbMakerApplication } from "../../../../cat-db-management/cat-dbMaker/cartas/dbMakerApplication"
 import { useSession, signIn, signOut } from "next-auth/react";
 import { unstable_getServerSession } from "next-auth/next"
-import authOptions from "../../auth/[...nextauth]"
+import authOptions from "../../../../utils/auth/options"
 
 
 let dbMakerApplication = new DbMakerApplication();
@@ -11,20 +11,22 @@ let dbMakerApplication = new DbMakerApplication();
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
     const session = await unstable_getServerSession(req, res, authOptions)
 
-    try{
-        if(session){
-            if(req.method === 'PUT')
-            {
+    try {
+        if (session) {
+            if (req.method === 'PUT') {
                 await dbMakerApplication.editSyllabusField(req.body)
                 res.status(200).json("Sucessful edit operation, or was it? thuuuuuum")
             }
-            else{
+            else {
                 res.status(400).json("Este endpoint es solo para solicitudes PUT para editar campos de templetes")
             }
+        }else {
+            res.status(401).json("Unathourized access.")
         }
-        }
- 
-    catch(err){
+    
+    }
+
+    catch (err) {
         res.status(500).json(err.message)
     }
 }
