+++ cat/pages/api/dbMakerAPI/cartas/getFields.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next"
 import { DbMakerApplication } from "../../../../cat-db-management/cat-dbMaker/cartas/dbMakerApplication"
 import { unstable_getServerSession } from "next-auth/next"
-import authOptions from "../../auth/[...nextauth]"
+import authOptions from "../../../../utils/auth/options"
 
 
 let dbMakerApplication = new DbMakerApplication();
@@ -19,7 +19,10 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
             else{
                 res.status(400).json("Este endpoint es solo para solicitudes GET para obtener los campos de templete")
             }
+        }else {
+            res.status(401).json("Unathourized access.")
         }
+    
     } catch (error) {
         res.status(500).json(error.message);
