+++ cat/pages/api/dbMakerAPI/templetes/getFields.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next"
 import { DbMakerApplication } from "../../../../cat-db-management/cat-dbMaker/templetes/dbMakerApplication"
 import { unstable_getServerSession } from "next-auth/next"
-import authOptions from "../../auth/[...nextauth]"
+import authOptions from "../../../../utils/auth/options"
 
 
 let dbMakerApplication = new DbMakerApplication();
@@ -19,6 +19,8 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
         else {
             res.status(400).json("Este endpoint es solo para solicitudes GET para obtener los campos de templete")
         }
+    } else {
+        res.status(401).json("Unathourized access.")
     }
 
 }
